api key changes

+++ TedsProject/Controllers/DbControllerController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Primitives;
 using TedsProject.Interfaces;
 using TedsProject.Models;
 
@@ -15,11 +16,13 @@ namespace TedsProject.Controllers
     {
 
         private readonly IDataService _dataService;
+        private readonly IKeysService _keysService;
         private readonly ILogger<DbControllerController> _logger;
 
-        public DbControllerController(IDataService dataService, ILogger<DbControllerController> logger)
+        public DbControllerController(IDataService dataService, IKeysService keysService, ILogger<DbControllerController> logger)
         {
             _dataService = dataService;
+            _keysService = keysService;
             _logger = logger;
 
         }
@@ -30,15 +33,23 @@ namespace TedsProject.Controllers
             return Response(_dataService.CreateTable());
         }
 
-        [HttpGet("all-crossings/{appKey}")]
-        public async Task<IActionResult> GetAll(string appkey)
+        [HttpGet("all-crossings")]
+        public async Task<IActionResult> GetAll()
         {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if(!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.GetAll());
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(string id)
         {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.GetById(id));
         }
 
@@ -51,33 +62,76 @@ namespace TedsProject.Controllers
         [HttpPut]
         public async Task<IActionResult> UPdateCrossing([FromBody] CrossingsModel model)
         {
-            
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.SaveItem(model));
         }
         
         [HttpDelete("{key}")]
         public async Task<IActionResult> DeleteCrossing(string key)
         {
-            
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.DeleteCrossing(key));
         }
 
-        [HttpGet("search/{lat}/{lng}/{key}"), AllowAnonymous]
+        [HttpGet("search/{lat}/{lng}/{key}")]
         public async Task<IActionResult> SearchCrossing(decimal lat, decimal lng, string key)
         {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.SearchBYLatLang(lat, lng, key));
         }
+        
 
         [HttpGet("get-gate-status/{id}")]
         public async Task<IActionResult> GetGateStatus(string id)
         {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
             return Response(await _dataService.GetGateStatus(id));
         }
 
         [HttpPost("update-gate-status/{id}")]
         public async Task<IActionResult> UpdateGateStatus([FromBody] GateStatus isOpen, string id)
         {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
+
             return Response(await _dataService.UpdateGateStatus(isOpen.isOpen, id));
+        }
+        
+        
+        
+        [HttpGet("get-gate-status/{country}/{tcnumber}")]
+        public async Task<IActionResult> GetGateStatusByTCNumber(string country, string tcnumber)
+        {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
+            return Response(await _dataService.GetGateStatusByTCNumber(country, tcnumber));
+        }
+
+        [HttpPost("update-gate-status/{country}/{tcnumber}")]
+        public async Task<IActionResult> UpdateGateStatusByTCNumber([FromBody] GateStatus isOpen, string country, string tcnumber)
+        {
+            var isKeyValid = await _keysService.ValidateAppKey(GetAppKey);
+            if (!isKeyValid)
+                return Response(errorMessage: "Invalid Api Key");
+
+
+            return Response(await _dataService.UpdateGateStatusByTCNumber(isOpen.isOpen, country, tcnumber));
         }
     }
